BAD SMELL Primitive Obsession: In many places there are units of sugar, milk, coffee, chocolate, and boullion stored as seperate ints. This poses many problems. First of all, it's messy and hard to keep track of everything. Second, the names of these ingredients are hardcoded and printed out in many places. So if the ingredients ever change it would be a huge hassle to fix all the code.
REFACTORING
Extract Class & Symbolic Constants: To fix this we made an ingredient class which stores the name and amount of an ingredient. We also made a String[] constant that stores the names of all the ingredients in one place, making it very simple to add/remove/edit ingredients. We added an Ingredient[] to the vending machine class to store the inventory, and initialize it in the vending machine constructor to have the names of all the ingredients in our String[] constant. Lastly, we made an Ingredient[] for the Recipe class to store the amount of each ingredient needed for the recipe.